+++ docs/changelog.md
@@ -4,6 +4,7 @@
 - Adds label support to lines with automatic angle adjustment
 - Plots the Moon's precise phase / illumination when `show_phase=True`
 - Separate collision handlers for points, areas, and paths
+- Refactored some models to expect an observer instance instead of lat/lon/dt
 
 
 ## v0.19.x